fix: address Codex review — fold theme tests into existing suite, fix guard test bugs
- Move theme-colors.test.ts assertions into existing terminal-theme.test.ts to
  avoid a generic feature test file (Codex P2)
- Fix chalk-named-color-guard.test.ts:
  - Replace require('fs') with ESM import (readdirSync) (P1)
  - Fix over-broad comment-skip logic — use proper inBlockComment state
    tracker instead of line.includes('*') which silently skipped valid
    code lines containing asterisks (P1)
  - Remove dead code (const entries = [['dir', '']]) (P2)
