+++ apps/kimi-code/test/tui/chalk-named-color-guard.test.ts
@@ -1,4 +1,4 @@
-import { readFileSync, statSync } from 'node:fs';
+import { readFileSync, readdirSync } from 'node:fs';
 import { join, relative } from 'node:path';
 
 import { describe, expect, it } from 'vitest';
@@ -17,9 +17,8 @@ const CHALK_NAMED_PATTERN = new RegExp(
 );
 
 function walk(dir: string, files: string[] = []): string[] {
-  const entries = [['dir', '']];
   try {
-    for (const entry of require('fs').readdirSync(dir, { withFileTypes: true })) {
+    for (const entry of readdirSync(dir, { withFileTypes: true })) {
       const p = join(dir, entry.name);
       if (entry.isDirectory()) {
         walk(p, files);
@@ -35,18 +34,31 @@ describe('chalk named color guard', () => {
   it('forbids chalk named colors in production source code', () => {
     const offenders: { file: string; line: number; snippet: string }[] = [];
     const files = walk(SRC_ROOT);
+    let inBlockComment = false;
     for (const file of files) {
       const content = readFileSync(file, 'utf8');
       const lines = content.split('\n');
       for (let i = 0; i < lines.length; i++) {
         const line = lines[i] ?? '';
-        if (line.trimStart().startsWith('//') || line.trimStart().startsWith('*')) continue;
-        if (line.includes('*')) continue;
+        const trimmed = line.trimStart();
+
+        if (inBlockComment) {
+          if (trimmed.includes('*/')) inBlockComment = false;
+          continue;
+        }
+
+        if (trimmed.startsWith('/*')) {
+          if (!trimmed.includes('*/')) inBlockComment = true;
+          continue;
+        }
+
+        if (trimmed.startsWith('//') || trimmed.startsWith('*')) continue;
+
         CHALK_NAMED_PATTERN.lastIndex = 0;
         const m = CHALK_NAMED_PATTERN.exec(line);
         if (m) {
           offenders.push({
-            file: relative(join(__dirname, '..', '..', 'src'), file),
+            file: relative(SRC_ROOT, file),
             line: i + 1,
             snippet: line.trim(),
           });
